pref:
优化商品草稿/商品模版，sku和批发规则保存规则。

BREAKING CHANGE: li_wholesale 表 增加 template_id(商品模版id)字段

+++ framework/src/main/java/cn/lili/modules/goods/sku/render/impl/WholesaleSaleModelRenderImpl.java
@@ -55,7 +55,9 @@ public class WholesaleSaleModelRenderImpl implements SalesModelRender {
             skus.setPrice(collect.get(0).getPrice());
             skus.setCost(collect.get(0).getPrice());
         }
-        wholesaleService.removeByGoodsId(goodsSkus.get(0).getGoodsId());
+        if (Boolean.FALSE.equals(goodsOperationDTO.getGoodsTemplateFlag())) {
+            wholesaleService.removeByGoodsId(goodsSkus.get(0).getGoodsId());
+        }
         wholesaleService.saveOrUpdateBatch(collect);
     }
